chore: create more page/header variations
