+++ apps/meteor/client/components/Page/PageBlock.tsx
@@ -1,0 +1,11 @@
+import type { Box } from '@rocket.chat/fuselage';
+import type { ComponentProps } from 'react';
+import React, { forwardRef } from 'react';
+
+import PageContent from './PageContent';
+
+const PageBlock = forwardRef<HTMLElement, ComponentProps<typeof Box>>(function PageBlock(props, ref) {
+	return <PageContent borderBlockEndColor='transparent' {...props} pb={16} ref={ref} borderBlockEndWidth='default' height='auto' />;
+});
+
+export default PageBlock;
