+++ apps/meteor/client/components/Page/PageBlockWithBorder.tsx
@@ -2,21 +2,12 @@ import type { Box } from '@rocket.chat/fuselage';
 import type { ComponentProps } from 'react';
 import React, { forwardRef, useContext } from 'react';
 
-import PageContent from './PageContent';
+import PageBlock from './PageBlock';
 import PageContext from './PageContext';
 
 const PageBlockWithBorder = forwardRef<HTMLElement, ComponentProps<typeof Box>>(function PageBlockWithBorder(props, ref) {
 	const [border] = useContext(PageContext);
-	return (
-		<PageContent
-			{...props}
-			pb={16}
-			ref={ref}
-			borderBlockEndColor={border ? 'extra-light' : 'transparent'}
-			borderBlockEndWidth='default'
-			height='auto'
-		/>
-	);
+	return <PageBlock ref={ref} {...props} borderBlockEndColor={border ? 'extra-light' : 'transparent'} />;
 });
 
 export default PageBlockWithBorder;
