+++ apps/meteor/client/components/Page/PageHeaderNoShadow.tsx
@@ -1,0 +1,50 @@
+import { Box, IconButton } from '@rocket.chat/fuselage';
+import { useDocumentTitle } from '@rocket.chat/ui-client';
+import { useLayout, useTranslation } from '@rocket.chat/ui-contexts';
+import type { ComponentPropsWithoutRef, ReactNode } from 'react';
+import React from 'react';
+
+import { HeaderToolbar } from '../Header';
+import SidebarToggler from '../SidebarToggler';
+
+type PageHeaderProps = {
+	title: ReactNode;
+	onClickBack?: () => void;
+	borderBlockEndColor?: string;
+} & Omit<ComponentPropsWithoutRef<typeof Box>, 'title'>;
+
+const PageHeaderNoShadow = ({ children = undefined, title, onClickBack, ...props }: PageHeaderProps) => {
+	const t = useTranslation();
+
+	const { isMobile } = useLayout();
+
+	useDocumentTitle(typeof title === 'string' ? title : undefined);
+
+	return (
+		<Box is='header' borderBlockEndWidth='default' borderBlockEndColor='transparent' {...props}>
+			<Box
+				height='100%'
+				marginInline={24}
+				minHeight='x64'
+				display='flex'
+				flexDirection='row'
+				flexWrap='wrap'
+				alignItems='center'
+				color='default'
+			>
+				{isMobile && (
+					<HeaderToolbar>
+						<SidebarToggler />
+					</HeaderToolbar>
+				)}
+				{onClickBack && <IconButton small mie={8} icon='arrow-back' onClick={onClickBack} title={t('Back')} />}
+				<Box is='h1' fontScale='h2' flexGrow={1} data-qa-type='PageHeader-title'>
+					{title}
+				</Box>
+				{children}
+			</Box>
+		</Box>
+	);
+};
+
+export default PageHeaderNoShadow;
